% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precis.R
\name{precis}
\alias{precis}
\title{Display a succinct summary of a data frame}
\usage{
precis(x, ..., width = 60)
}
\arguments{
\item{x}{A tabular structure like a data frame.}

\item{...}{Other arguments potentially used by methods}

\item{width}{Maximum width for each line.}
}
\description{
Display a succinct summary of a data frame
}
\examples{
precis(mtcars)
}

