% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precis_v.R
\name{precis_v}
\alias{precis_v}
\title{Succint summary of a vector}
\usage{
precis_v(x, ..., width = 60)
}
\arguments{
\item{x}{A vector}

\item{...}{Other arguments potentially used by methods}

\item{width}{Maximum width for each line.}
}
\value{
A character vector, with one entry for each line.
}
\description{
Succint summary of a vector
}
\examples{
precis_v(mtcars[[1]])
}
\keyword{internal}

